// frontend/src/shared/api/generated/models/profile-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pied Piper Freelance API
 * API для биржи фриланса (MVP)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProfileSkillDetailedDTO } from './profile-skill-detailed-dto';

export interface ProfileDTO {
    'accountId'?: number;
    'email'?: string;
    'role'?: string;
    'accountStatus'?: string;
    'lastLogin'?: string;
    'accountCreatedAt'?: string;
    'profileId'?: number;
    'firstName'?: string;
    'lastName'?: string;
    'bio'?: string;
    'hourlyRate'?: number;
    'avatarUrl'?: string;
    'profileUpdatedAt'?: string;
    'rating'?: number;
    'balance'?: number;
    'skills'?: Array<ProfileSkillDetailedDTO>;
}